requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
google-genai
elevenlabs
python-dotenv